+++ tests/shakedown/shakedown/clients/metronome.py
@@ -2,8 +2,10 @@ import json
 
 from six.moves import urllib
 
-from dcos import config, cosmos, http, packagemanager, rpcclient, util
-from dcos.errors import DCOSException
+from dcos import config, util
+from shakedown import http
+from shakedown.clients import cosmos, packagemanager, rpcclient
+from shakedown.errors import DCOSException
 
 logger = util.get_logger(__name__)
 
@@ -19,7 +21,7 @@ def create_client(toml_config=None):
     :param toml_config: configuration dictionary
     :type toml_config: config.Toml
     :returns: Metronome client
-    :rtype: dcos.metronome.Client
+    :rtype: shakedown.clients.metronome.Client
     """
 
     if toml_config is None:
