+++ tests/shakedown/shakedown/clients/packagemanager.py
@@ -5,11 +5,12 @@ import functools
 import six
 from six.moves import urllib
 
-from dcos import cosmos, emitting, util
-from dcos.errors import (DCOSAuthenticationException,
-                         DCOSAuthorizationException, DCOSBadRequest,
-                         DCOSConnectionError, DCOSException, DCOSHTTPException,
-                         DefaultError)
+from dcos import emitting, util
+from shakedown.clients import cosmos
+from shakedown.errors import (DCOSAuthenticationException,
+                              DCOSAuthorizationException, DCOSBadRequest,
+                              DCOSConnectionError, DCOSException, DCOSHTTPException,
+                              DefaultError)
 
 logger = util.get_logger(__name__)
 emitter = emitting.FlatEmitter()
