+++ tests/shakedown/shakedown/dcos/helpers.py
@@ -2,8 +2,6 @@ import os
 import paramiko
 import scp
 
-from dcos import http
-
 import itertools
 import shakedown
