+++ tests/shakedown/shakedown/dcos/marathon.py
@@ -1,8 +1,9 @@
 from distutils.version import LooseVersion
-from dcos import marathon, config
+from shakedown.clients import marathon
+from dcos import config
+from shakedown import *
 from shakedown.dcos.spinner import *
 from shakedown.dcos.service import service_available_predicate
-from shakedown import *
 from six.moves import urllib
 
 import pytest
