+++ tests/system/dcos_service_marathon_tests.py
@@ -6,8 +6,8 @@ import retrying
 import time
 
 from datetime import timedelta
-from dcos import marathon
-from dcos.errors import DCOSUnprocessableException
+from shakedown.clients import marathon
+from shakedown.errors import DCOSUnprocessableException
 
 
 @retrying.retry(wait_fixed=1000, stop_max_attempt_number=16, retry_on_exception=common.ignore_exception)
